requirements.txt
Django~=5.2.0
django-auth-ldap~=5.1.0
asgiref~=3.8.1
sqlparse~=0.5.3
# 基础包
mysqlclient~=2.2.7
orjson~=3.10.16
lxml~=5.3.2
typing_extensions~=4.13.2
# 必须whl安装的包
python-ldap~=3.4.4
# 项目包
django-environ~=0.12.0
django-ninja~=1.4.1
pydantic~=2.11.3
pydantic-core~=2.33.1
django-ninja-extra~=0.30.0
django-ninja-jwt~=5.3.7
django-tinymce~=4.1.0
PyJWT~=2.10.1
cryptography~=44.0.2
ninja-schema~=0.14.2
user-agents~=2.2.0
numpy~=2.2.5
docxtpl~=0.19.1
python-docx~=1.1.2
Jinja2~=3.1.6
django-redis~=5.4.0
pandas~=2.2.3
beautifulsoup4~=4.13.4
python-dateutil~=2.9.0.post0
redis~=5.2.1
waitress~=3.0.2
django-shortuuidfield~=0.1.3
lizard~=1.17.27
Pygments~=2.19.1